Add "Terms and Conditions" and "Privacy Policy" to the landing page
These are apparently required in order to get ads accepted by Reddit. Since we're a research project and not a company, this change mentions both of the exact phrases that are apparently required but points people to the Informed Consent to Participate In Research Agreement (i.e. the PDF of what we show on the later consent screen.)