Log errors but redirect anyway when storing tracked clicks fails
Issues with click tracking are infrequent but do happen, and when they do, it has a negative impact on user experience that may discourage people from being active users because that prevents people from actually reading the news.

In those cases, let's log an error message we can alert on but redirect to the linked article anyway. That will minimize user experience disruptions while still providing us a way to know there's a problem.